Updated README to make it clearer for beginners, renamed coinflip file to coinFlip to keep consistency, and renamed parking command to garage
