+++ README.md
@@ -3,8 +3,12 @@ Discord bot created to serve the KnightHacks discord. Feel free to join us [Here
 
 ## Setup
 1. Clone this repository on your environment.
-2. Install the latest nodejs and npm.
-3. Visit the directory and run `npm install`
+2. Install the latest nodejs and npm:
+    a. https://nodejs.org/en/
+    b. Instructions for Mac: https://treehouse.github.io/installation-guides/mac/node-mac.html
+    c. Instructions for Windows: https://www.guru99.com/download-install-node-js.html
+    d. Instructions for Linux: https://blog.teamtreehouse.com/install-node-js-npm-linux
+3. Visit the project's main directory and run `npm install`
 4. Create a `constants/token.json` file and format and insert your discord bot token properly. (Refer to token.json section)
 5. Run `npm start` to start the discord bot.
 
@@ -21,6 +25,7 @@ Place the token.json in the `constants/` directory.
 1. Run your code through ESLint first. (Refer to ESLint section)
 2. Read over your comments and variable names and check for speelign erorrs.
 3. Check for compilation errors.
+4. Make sure your code is consistent with the rest of the project and with other commands (i.e. diceRoll.js, not diceroll.js).
 
 ## ESLint
 > What is ESLint?
@@ -44,3 +49,9 @@ ESLint is a way to create a uniform code base throughout the project. It is bene
     > https://docs.npmjs.com/misc/coding-style.html
 
     Please check at `.eslintrc.js` for more rules that are not standardized by the standards above.
+
+
+### Troubleshooting
+1. If the bot is not starting, try deleting your node_modules folder and your package-lock.json file (both located in the project's main directory), and running npm install again.
+2. Feel free to reach out to Lancelot devs on Discord for questions and troubleshooting.
+3. Let us know if there's anything that needs to be added to this section!
